chore(suite-native): move setTemporaryRememberedDeviceThunk to firmware package
