+++ suite-native/firmware/package.json
@@ -15,6 +15,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@shopify/react-native-skia": "2.2.12",
         "@suite-common/firmware": "workspace:*",
+        "@suite-common/redux-utils": "workspace:*",
         "@suite-common/icons": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
