+++ suite-native/firmware/src/components/FirmwareInstallationScreenContent.tsx
@@ -7,11 +7,7 @@ import { useKeepAwake } from 'expo-keep-awake';
 
 import { firmwareActions } from '@suite-common/firmware';
 import { Box, Button, VStack, useBottomSheetModal } from '@suite-native/atoms';
-import {
-    ConfirmOnTrezorWrapper,
-    setTemporaryRememberedDeviceThunk,
-    useConfirmOnTrezorController,
-} from '@suite-native/device';
+import { ConfirmOnTrezorWrapper, useConfirmOnTrezorController } from '@suite-native/device';
 import { Translation } from '@suite-native/intl';
 import { SUITE_MOBILE_SUPPORT_URL, useOpenLink } from '@suite-native/link';
 import { DynamicScreenHeader } from '@suite-native/navigation';
@@ -19,6 +15,7 @@ import { reportSecurityCheck } from '@suite-native/sentry';
 import TrezorConnect from '@trezor/connect';
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 
+import { setTemporaryRememberedDeviceThunk } from '../firmwareThunks';
 import { DoNotCloseAppBottomSheetTrigger } from './DoNotCloseAppBottomSheetTrigger';
 import { FirmwareInstallationProgressTitles } from './FirmwareInstallationProgressTitles';
 import { MayBeStuckedBottomSheet } from './MayBeStuckedBottomSheet';
